changed everything to use floats instead of double to match given io code

+++ src/cs3500/animator/model/interfaces/Drawable.java
@@ -41,7 +41,7 @@ public interface Drawable {
    * @param g the green value
    * @param b the blue value
    */
-  void setColor(int r, int g, int b);
+  void setColor(float r, float g, float b);
 
   /**
    * Changes the color of this object.
